\name{cerardat_app}
\alias{cerardat_app}
\title{
  Launch the shiny application.
}
\description{
  see \link[SPARTAAS]{cerardat}.
  }
\usage{
  cerardat_app()
}

\value{No return value}

\author{
  A. COULON

  L. BELLANGER

  P. HUSI
}



\examples{
  ##---- Should be DIRECTLY executable !! ----
  ##-- ==>  Define data, use random,
  ##--	or do  help(data=index)  for the standard data sets.
  \donttest{
    library(SPARTAAS)
    if(interactive()){
      cerardat_app()
    }
  }

}

