\name{mapclust_app}
\alias{mapclust_app}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Shiny application for Mapclust method
}
\description{
see \link[SPARTAAS]{mapclust}
You can also check the wiki on the application.
}
\usage{
mapclust_app()
}


\value{No return value}

\author{
A. COULON

L. BELLANGER

P. HUSI
}

\examples{
#open application
#library(SPARTAAS)
#if(interactive()){
  #mapclust_app()
#}
}

