\name{serio_app}
\alias{serio_app}
\title{
Launch the shiny application.
}
\description{
see \link[SPARTAAS]{seriograph}.
You can also check the wiki on the application.}
\usage{
serio_app()
}

\value{No return value}

\author{
A. COULON

L. BELLANGER

P. HUSI
}



\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
\donttest{
library(SPARTAAS)
if(interactive()){
  serio_app()
}
}

}

