\name{SPARTAAS-package}
\alias{SPARTAAS-package}
\alias{SPARTAAS}
\docType{package}
\title{
SPARTAAS
}
\description{
\strong{S}tatistical \strong{Pa}ttern \strong{R}ecognition and da\strong{T}ing using \strong{A}rcheological \strong{A}rtefacts assemblage\strong{S}:
hclustcompro(perioclust): compromised hierarchical agglomerative clustering method.
}
\details{
\packageDescription{SPARTAAS}
}
\author{
\packageAuthor{SPARTAAS}

Maintainer: \packageMaintainer{SPARTAAS}
}
\references{
Bellanger L., Coulon A., Husi P., 2020 – Perioclust: a new Hierarchical agglomerative clustering method including temporal or spatial ordering constraints. Springer Series, Studies in Classification, Data Analysis, and Knowledge Organization. <doi: 10.1007/978-3-030-60104-1>

Bellanger L., Husi P., Laghzali Y. (2015). Spatial statistic analysis of dating using pottery: an aid to the characterization of cultural areas in West Central France. In : Traviglia A. ed., Across Space and Time, Proceedings of the 41th International Conference on Computer Applications and Quantitative Methods in Archaeology (CAA-2013), Perth (Australie), Amsterdam University Press : 276-282.
}

